excluded functionality from index
